+++ src/Trash/CompositionRoot.cs
@@ -95,8 +95,10 @@ namespace Trash
                 .As<IConfigurationProvider>()
                 .SingleInstance();
 
-            builder.Register(c => c.Resolve<IConfigurationProvider>().ActiveConfiguration)
-                .As<IServiceConfiguration>();
+            // note: Do not allow consumers to resolve IServiceConfiguration directly; if this gets cached
+            // they end up using the wrong configuration when multiple instances are used.
+            // builder.Register(c => c.Resolve<IConfigurationProvider>().ActiveConfiguration)
+            // .As<IServiceConfiguration>();
         }
 
         private static void CommandRegistrations(ContainerBuilder builder)